requirements.txt
Cython==0.29.21
librosa==0.9.2
matplotlib==3.3.1
scikit-learn==1.0.2
scipy
numpy==1.22
tensorboard
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
unidecode
pyopenjtalk-prebuilt
jamo
pypinyin
jieba
protobuf
cn2an
inflect
eng_to_ipa
ko_pron
indic_transliteration==2.3.37
num_thai==0.0.5
opencc==1.1.1
demucs
git+https://github.com/openai/whisper.git 
gradio
pymlask @ git+https://github.com/ikegami-yukino/pymlask@227ce3b42bba9ede71165ec10735ed9d18f6aa37
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyYAML==6.0.2
mecab==0.996.3
openai==1.47.1
pillow==10.4.0
